+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Fix rounding for progression percentage (#218)
 -   Fix error in console due to bad markup (#213)
 -   Removed warning in the console (#221)
+-   Fix infinite load on the train tasks compute plan on first load (#225)
 
 ### Changed